hotel-reservation/services/hotel/requirements.txt
fastapi
strawberry-graphql[fastapi]
uvicorn